add tests and address comment
