+++ src/bin/tikv-server.rs
@@ -64,6 +64,7 @@ use tikv::util::collections::HashMap;
 use tikv::util::logger::{self, StderrLogger};
 use tikv::util::file_log::RotatingFileLogger;
 use tikv::util::transport::SendCh;
+use tikv::util::properties::UserPropertiesCollectorFactory;
 use tikv::server::{DEFAULT_LISTENING_ADDR, DEFAULT_CLUSTER_ID, Server, Node, Config,
                    create_raft_storage};
 use tikv::server::transport::ServerRaftStoreRouter;
@@ -580,7 +581,7 @@ fn get_rocksdb_write_cf_option(config: &toml::Value, total_mem: u64) -> RocksdbO
     // Create prefix bloom filter for memtable.
     opts.set_memtable_prefix_bloom_size_ratio(0.1 as f64);
     // Collects user defined properties.
-    let f = Box::new(rocksdb_util::UserPropertiesCollectorFactory::default());
+    let f = Box::new(UserPropertiesCollectorFactory::default());
     opts.add_table_properties_collector_factory("tikv.user-properties-collector", f);
     opts
 }
